requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sendgrid
Jinja2